+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v1.1.0
+## new
+- chiyo_analytics.toml 新增配置项：
+    - redis, postgres, clickhouse 新增： deploy.single_server_docker 选项。允许配置绑定的本机端口，挂载的 volume 配置等等
+    - collector, api 的 cors_allowed_origins 允许使用通配符。
+
 # v1.0.1
 ## fix
 - 修复了 single server docker 部署的管理 CLI 不尊重安装 CLI 的 --dest 参数的问题